use big decimal

+++ server/src/main/java/com/zafira/safe_zafira/vehicle/model/Vehicle.java
@@ -1,12 +1,14 @@
 package com.zafira.safe_zafira.vehicle.model;
 
+import java.math.BigDecimal;
+
 public record Vehicle(
         Long id,
         String vehicleNo,
         String vin,
         String make,
         String model,
-        Double batteryVoltage
+        BigDecimal batteryVoltage
 ) {
 }
